Split review bot diagnostic presenters
